+++ wp-content/plugins/cleantalk-spam-protect/cleantalk.php
@@ -3,15 +3,15 @@
   Plugin Name: Anti-Spam by CleanTalk
   Plugin URI: http://cleantalk.org
   Description: Max power, all-in-one, no Captcha, premium anti-spam plugin. No comment spam, no registration spam, no contact spam, protects any WordPress forms.
-  Version: 5.97
+  Version: 5.98
   Author: СleanTalk <welcome@cleantalk.org>
   Author URI: http://cleantalk.org
 */
 
 $cleantalk_executed = false;
 
-define('APBCT_VERSION', '5.97');
-define('APBCT_AGENT',   'wordpress-597');
+define('APBCT_VERSION', '5.98');
+define('APBCT_AGENT',   'wordpress-598');
 
 define('CLEANTALK_REMOTE_CALL_SLEEP', 10); // Minimum time between remote call
 define('APBCT_CASERT_PATH', file_exists(ABSPATH . WPINC . '/certificates/ca-bundle.crt')