+++ wp-content/plugins/social-polls-by-opinionstage/assets/content-popup/app/components/new-widget.js
@@ -1,5 +1,0 @@
-import Vue from 'vue'
-
-export default Vue.component('new-widget', {
-  template: '#opinionstage-new-widget',
-})