+++ wp-content/plugins/social-polls-by-opinionstage/opinionstage-polls.php
@@ -1,11 +1,11 @@
 <?php
 /*
-Plugin Name: Poll, Survey, Quiz, Slideshow, Form, Story & Landing Page
-Plugin URI: http://www.opinionstage.com
+Plugin Name: Poll, Survey, Quiz, Slideshow, Form & Story Article
+Plugin URI: https://www.opinionstage.com
 Description: Add a highly engaging poll, survey, quiz or contact form builder to your site. You can add the poll, survey, quiz or form to any post/page or to the sidebar.
-Version: 19.3.3
+Version: 19.3.5
 Author: OpinionStage.com
-Author URI: http://www.opinionstage.com
+Author URI: https://www.opinionstage.com
 Text Domain: social-polls-by-opinionstage
 */
 
@@ -24,7 +24,7 @@ if ( defined('WP_DEBUG') && true === WP_DEBUG ) {
 	}
 }
 
-define('OPINIONSTAGE_WIDGET_VERSION', '19.3.3');
+define('OPINIONSTAGE_WIDGET_VERSION', '19.3.5');
 
 define('OPINIONSTAGE_TEXT_DOMAIN', 'social-polls-by-opinionstage');